+++ spec/support/policies/post_policy.rb
@@ -33,4 +33,17 @@ class PostPolicy < BasePolicy
   def permitted_attributes_for_revise
     [:body]
   end
+
+  def expected_attributes_for_action(action_name)
+    case action_name.to_sym
+    when :revise
+      [:body]
+    else
+      if post.user == user
+        %i[title votes]
+      else
+        [:votes]
+      end
+    end
+  end
 end